requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
discord.py==1.7.3
Django==4.0.5
django-environ==0.8.1
django-shortcuts==1.6
djangorestframework==3.13.1
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
yarl==1.7.2
